feat: getFileIdFor functions
