+++ Sources/Hiero/EntityId.swift
@@ -231,63 +231,3 @@ internal enum PartialEntityId<S: StringProtocol> {
         }
     }
 }
-
-// fixme(sr): How do DRY?
-
-/// The unique identifier for a file on Hiero.
-public struct FileId: EntityId, ValidateChecksums, Sendable {
-    public init(shard: UInt64 = 0, realm: UInt64 = 0, num: UInt64, checksum: Checksum?) {
-        self.shard = shard
-        self.realm = realm
-        self.num = num
-        self.checksum = checksum
-    }
-
-    public init(shard: UInt64 = 0, realm: UInt64 = 0, num: UInt64) {
-        self.init(shard: shard, realm: realm, num: num, checksum: nil)
-    }
-
-    public let shard: UInt64
-    public let realm: UInt64
-
-    /// The file number.
-    public let num: UInt64
-
-    public let checksum: Checksum?
-
-    public static let addressBook: FileId = 102
-    public static let feeSchedule: FileId = 111
-    public static let exchangeRates: FileId = 112
-
-    public static func fromBytes(_ bytes: Data) throws -> Self {
-        try Self(protobufBytes: bytes)
-    }
-
-    public func toBytes() -> Data {
-        toProtobufBytes()
-    }
-
-    internal func validateChecksums(on ledgerId: LedgerId) throws {
-        try helper.validateChecksum(on: ledgerId)
-    }
-}
-
-extension FileId: ProtobufCodable {
-    internal typealias Protobuf = HieroProtobufs.Proto_FileID
-
-    internal init(protobuf proto: Protobuf) {
-        self.init(
-            shard: UInt64(proto.shardNum),
-            realm: UInt64(proto.realmNum),
-            num: UInt64(proto.fileNum)
-        )
-    }
-
-    internal func toProtobuf() -> Protobuf {
-        .with { proto in
-            proto.shardNum = Int64(shard)
-            proto.realmNum = Int64(realm)
-            proto.fileNum = Int64(num)
-        }
-    }
-}
